Feat/Fix - The Delete data function now properly deletes data and/or alerts the user that the email was not found.
Several JS event listeners had to be wrapped in "checks" as apparently, when using the same JS file across different pages, event listeners not on that page break the js file altogether. By wrapping them in an if statement that simply checks if the variable that acts as the query selector exists, this issue can be avoided.